backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
# Auth
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# Scheduler
apscheduler>=3.10.0
# ML (Unnati's models)
xgboost>=2.0.0
lightgbm>=4.0.0
shap>=0.44.0
duckdb>=0.10.0
holidays>=0.46
joblib>=1.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
